Remove unused functions, rearrange
